requirements.txt
flask
face_recognition
dlib
opencv-python
numpy
mysql-connector-python
twilio

